@@ -1,8 +1,8 @@
 [package]
 name         = "openweathermap_client"
 description  = "A client for querying the weather from the free OpenWeatherMap API."
-version      = "0.4.0"
-rust-version = "1.71"
+version      = "0.5.0"
+rust-version = "1.82"
 edition      = "2021"
 authors      = [ "evaneaston" ]
 homepage     = "https://github.com/evaneaston/openweathermap"
